test: add more test cases and extract to a separate class
