+++ spark-3/src/test/scala/org/neo4j/spark/DataSourceWriterNeo4jSkipNullKeysTSE.scala
@@ -1,0 +1,495 @@
+package org.neo4j.spark
+
+import org.apache.spark.SparkException
+import org.apache.spark.sql.SaveMode
+import org.junit.Test
+import org.neo4j.Closeables.use
+
+class DataSourceWriterNeo4jSkipNullKeysTSE extends SparkConnectorScalaBaseTSE {
+
+  import ss.implicits._
+
+  @Test
+  def `fails to write nodes when mode is not Overwrite`(): Unit = {
+    val cities = Seq(
+      (Some(1), "Cherbourg en Cotentin"),
+      (Some(2), "London"),
+      (Some(3), "Malmö")
+    ).toDF("id", "city")
+
+    val caught = intercept[IllegalArgumentException] {
+      cities.write
+        .format(classOf[DataSource].getName)
+        .mode(SaveMode.Append)
+        .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+        .option("labels", ":City")
+        .option("node.keys", "id")
+        .option("skip.null.keys", "true")
+        .save()
+    }
+    assert(caught.getMessage contains "`skip.null.keys` works only with `mode` `Overwrite`")
+  }
+
+  @Test
+  def `fails to write nodes when key properties contain null values`(): Unit = {
+    val cities = Seq(
+      (Some(1), "Cherbourg en Cotentin"),
+      (Some(2), "London"),
+      (Some(3), "Malmö"),
+      (None, "Moon")
+    ).toDF("id", "city")
+
+    val caught = intercept[SparkException] {
+      cities.write
+        .format(classOf[DataSource].getName)
+        .mode(SaveMode.Overwrite)
+        .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+        .option("labels", ":City")
+        .option("node.keys", "id")
+        .option("schema.optimization.node.keys", "KEY")
+        .save()
+    }
+    assert(caught.getMessage contains "Cannot merge the following node because of null property value")
+  }
+
+  @Test
+  def `fails to write relationships when source node key properties contain null values`(): Unit = {
+    val caught = intercept[SparkException] {
+      val cities = Seq(
+        (Some(1), Some(2), "British Airways"),
+        (Some(2), Some(3), "Turkish Airlines"),
+        (None, Some(5), "Another Airline")
+      ).toDF("from", "to", "airline")
+
+      cities.write
+        .format(classOf[DataSource].getName)
+        .mode(SaveMode.Overwrite)
+        .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+        .option("relationship", "FLIES_TO")
+        .option("relationship.save.strategy", "keys")
+        .option("relationship.source.save.mode", "Overwrite")
+        .option("relationship.source.labels", ":City")
+        .option("relationship.source.node.keys", "from:id")
+        .option("relationship.target.save.mode", "Overwrite")
+        .option("relationship.target.labels", ":City")
+        .option("relationship.target.node.keys", "to:id")
+        .option("relationship.properties", "airline")
+        .option("schema.optimization.node.keys", "KEY")
+        .save()
+    }
+    assert(caught.getMessage contains "Cannot merge the following node because of null property value")
+  }
+
+  @Test
+  def `fails to write relationships when target node key properties contain null values`(): Unit = {
+    val caught = intercept[SparkException] {
+      val cities = Seq(
+        (Some(1), Some(2), "British Airways"),
+        (Some(2), Some(3), "Turkish Airlines"),
+        (Some(3), None, "Another Airline")
+      ).toDF("from", "to", "airline")
+
+      cities.write
+        .format(classOf[DataSource].getName)
+        .mode(SaveMode.Overwrite)
+        .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+        .option("relationship", "FLIES_TO")
+        .option("relationship.save.strategy", "keys")
+        .option("relationship.source.save.mode", "Overwrite")
+        .option("relationship.source.labels", ":City")
+        .option("relationship.source.node.keys", "from:id")
+        .option("relationship.target.save.mode", "Overwrite")
+        .option("relationship.target.labels", ":City")
+        .option("relationship.target.node.keys", "to:id")
+        .option("relationship.properties", "airline")
+        .option("schema.optimization.node.keys", "KEY")
+        .save()
+    }
+    assert(caught.getMessage contains "Cannot merge the following node because of null property value")
+  }
+
+  @Test
+  def `fails to write relationships when relationship key properties contain null values`(): Unit = {
+    val caught = intercept[SparkException] {
+      val cities = Seq(
+        (Some(1), Some(2), Some("BA721"), "British Airways"),
+        (Some(2), Some(3), Some("TK211"), "Turkish Airlines"),
+        (Some(3), Some(4), None, "Another Airline")
+      ).toDF("from", "to", "flight", "airline")
+
+      cities.write
+        .format(classOf[DataSource].getName)
+        .mode(SaveMode.Overwrite)
+        .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+        .option("relationship", "FLIES_TO")
+        .option("relationship.save.strategy", "keys")
+        .option("relationship.source.save.mode", "Overwrite")
+        .option("relationship.source.labels", ":City")
+        .option("relationship.source.node.keys", "from:id")
+        .option("relationship.target.save.mode", "Overwrite")
+        .option("relationship.target.labels", ":City")
+        .option("relationship.target.node.keys", "to:id")
+        .option("relationship.keys", "flight")
+        .option("relationship.properties", "airline")
+        .option("schema.optimization.node.keys", "KEY")
+        .option("schema.optimization.relationship.keys", "KEY")
+        .save()
+    }
+    assert(caught.getMessage contains "Cannot merge the following relationship because of null property value")
+  }
+
+  @Test
+  def `skips nodes when key properties contain null values`(): Unit = {
+    val cities = Seq(
+      (Some(1), "Cherbourg en Cotentin"),
+      (Some(2), "London"),
+      (Some(3), "Malmö"),
+      (None, "Moon")
+    ).toDF("id", "city")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Overwrite)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("labels", ":City")
+      .option("node.keys", "id")
+      .option("schema.optimization.node.keys", "KEY")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val result = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(result == 3)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values`(): Unit = {
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(2), Some(3), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Overwrite)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Overwrite")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Overwrite")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("schema.optimization.node.keys", "KEY")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 3)
+
+      val citiesWithId5 = session.run("MATCH (n:City {id: 5}) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(citiesWithId5 == 0)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values when nodes are matched`()
+    : Unit = {
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      session.run("UNWIND [1,2,3,5] AS id CREATE (:City {id: id})").consume()
+    }
+
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(2), Some(3), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Overwrite)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Match")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Match")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 4)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values when nodes are appended`()
+    : Unit = {
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(3), Some(4), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Overwrite)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Append")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Append")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 4)
+
+      val citiesWithId5 = session.run("MATCH (n:City {id: 5}) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(citiesWithId5 == 0)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values with append mode`(): Unit = {
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(3), Some(4), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Append)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Overwrite")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Overwrite")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("schema.optimization.node.keys", "KEY")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 4)
+
+      val citiesWithId5 = session.run("MATCH (n:City {id: 5}) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(citiesWithId5 == 0)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values when nodes are matched with append mode`()
+    : Unit = {
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      session.run("UNWIND [1,2,3,5] AS id CREATE (:City {id: id})").consume()
+    }
+
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(2), Some(3), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Append)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Match")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Match")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 4)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when source or target node key properties contain null values when nodes are appended with append mode`()
+    : Unit = {
+    val cities = Seq(
+      (Some(1), Some(2), "British Airways"),
+      (Some(3), Some(4), "Turkish Airlines"),
+      (None, Some(5), "Another Airline"),
+      (Some(5), None, "Another Airline")
+    ).toDF("from", "to", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Append)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Append")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Append")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.properties", "airline")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 4)
+
+      val citiesWithId5 = session.run("MATCH (n:City {id: 5}) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(citiesWithId5 == 0)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+  @Test
+  def `skips relationships when relationship key properties contain null values`(): Unit = {
+    val cities = Seq(
+      (Some(1), Some(2), Some("BA721"), "British Airways"),
+      (Some(2), Some(3), Some("TK211"), "Turkish Airlines"),
+      (Some(3), Some(5), None, "Another Airline")
+    ).toDF("from", "to", "flight", "airline")
+
+    cities.write
+      .format(classOf[DataSource].getName)
+      .mode(SaveMode.Overwrite)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("relationship", "FLIES_TO")
+      .option("relationship.save.strategy", "keys")
+      .option("relationship.source.save.mode", "Overwrite")
+      .option("relationship.source.labels", ":City")
+      .option("relationship.source.node.keys", "from:id")
+      .option("relationship.target.save.mode", "Overwrite")
+      .option("relationship.target.labels", ":City")
+      .option("relationship.target.node.keys", "to:id")
+      .option("relationship.keys", "flight")
+      .option("relationship.properties", "airline")
+      .option("schema.optimization.node.keys", "KEY")
+      .option("schema.optimization.relationship.keys", "KEY")
+      .option("skip.null.keys", "true")
+      .save()
+
+    use(SparkConnectorScalaSuiteIT.driver.session()) { session =>
+      val cities = session.run("MATCH (n:City) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(cities == 3)
+
+      val citiesWithId5 = session.run("MATCH (n:City {id: 5}) RETURN count(n) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(citiesWithId5 == 0)
+
+      val flies = session.run("MATCH ()-[r:FLIES_TO]->() RETURN count(r) as count")
+        .single()
+        .get("count")
+        .asLong()
+      assert(flies == 2)
+    }
+  }
+
+}
